Convert eslint.conf to TS
